TableForm: Move TableForm from the core to the printing module.
