+++ sympy/__init__.py
@@ -69,7 +69,7 @@ from parsing import *
 #from physics import units
 from plotting import Plot, textplot
 from printing import pretty, pretty_print, pprint, pprint_use_unicode, \
-    pprint_try_use_unicode, print_gtk, print_tree, pager_print
+    pprint_try_use_unicode, print_gtk, print_tree, pager_print, TableForm
 from printing import ccode, fcode, latex, preview
 from printing import python, print_python, srepr, sstr, sstrrepr
 from interactive import init_session, init_printing
