+++ sympy/core/__init__.py
@@ -24,7 +24,7 @@ from function import Lambda, WildFunction, Derivative, diff, FunctionClass, \
 from sets import (Set, Interval, Union, EmptySet, FiniteSet, ProductSet,
         Intersection)
 from evalf import PrecisionExhausted, N
-from containers import Tuple, Dict, TableForm
+from containers import Tuple, Dict
 from exprtools import gcd_terms, factor_terms, factor_nc
 
 # expose singletons
